Move conditionally compiled code into their module
* Create the `auth::check` module, that exposes the `check_auth_for`
  function, which implementation is based on the `auth` compilation
  flags, like before.
  * Edit the arguments this function takes, by asking for connections
    directly instead of pools.
* Create the `auth::gen_token` module, that exposes the `gen_token_for`
  function, and that is only available with the `auth` compilation flag,
  as before.
* Create `player::auth` module, that exposes the `get_token` and
  `post_give_token` functions, which implementations are based on the
  `auth` compilation flags, like before.
* All this is to avoid having conditionally compiled imports in some
  modules, which is ugly. Some of them are still present, but they're
  acceptable.
