+++ crates/game_api/src/must.rs
@@ -1,15 +1,11 @@
 //! See [`records_lib::must`] module documentation.
 
 use actix_web::{FromRequest, HttpMessage, HttpRequest};
+use records_lib::RedisPool;
 use sea_orm::DbConn;
 use tracing_actix_web::RequestId;
 
-use crate::utils::ExtractDbConn;
-
-#[cfg(auth)]
-use crate::Res;
-#[cfg(auth)]
-use records_lib::Database;
+use crate::{Res, utils::ExtractDbConn};
 
 pub fn have_request_id(req: &HttpRequest) -> RequestId {
     *req.extensions()
@@ -17,17 +13,16 @@ pub fn have_request_id(req: &HttpRequest) -> RequestId {
         .expect("RequestId should be present")
 }
 
-#[cfg(auth)]
-pub fn have_db(req: &HttpRequest) -> Res<Database> {
-    req.app_data::<Database>()
-        .expect("Database app data should be present")
-        .clone()
-        .into()
+pub fn have_redis_pool(req: &HttpRequest) -> RedisPool {
+    <Res<RedisPool> as FromRequest>::extract(req)
+        .into_inner()
+        .unwrap()
+        .0
 }
 
 pub fn have_dbconn(req: &HttpRequest) -> DbConn {
     <ExtractDbConn as FromRequest>::extract(req)
         .into_inner()
-        .expect("ExtractDbConnection::from_request must not return an error")
+        .unwrap()
         .0
 }
